requirements.txt
openapi-spec-validator
ansicolors
jsonschema
json_schema_generator
